+++ gdx-jogo/src/com/minimine/utils/ruidos/OpenSimplex2.java
@@ -27,25 +27,25 @@ public final class OpenSimplex2 {
     // === CONSTANTES ===
 
     // 2D: fator de skew e de-skew do triangulo para o simplex
-    public static final double SKEW2 = 0.366025403784438646; // (sqrt(3) - 1) / 2
-    public static final double UNSKEW2 = 0.211324865405187117; // (3 - sqrt(3)) / 6
+    public static final float SKEW2 = 0.366025403784438646f; // (sqrt(3) - 1) / 2
+    public static final float UNSKEW2 = 0.211324865405187117f; // (3 - sqrt(3)) / 6
 
     // 2D: raio do kernel ao quadrado, OpenSimplex2S usa 9/4(maior que Simplex padrão 1/2)
     // raio maior -> transição mais suave entre contribuições de vertice
-    public static final double KERNEL2 = 0.5625; // (3/4)^2
+    public static final float KERNEL2 = 0.5625f; // (3/4)^2
 
     // normalização 2D, escala o resultado para ficar proximo de [-1, 1]
-    public static final double NORM2   = 47.0;
+    public static final float NORM2   = 47.0f;
 
     // 3D: rotação BCC(1/sqrt(3) - 1) / 3 e 1/sqrt(3) * 2/3
-    public static final double ROTATE3 = -0.16666666666666666; // (1/sqrt(3) - 1) / 3
-    public static final double FALLBACK3 = 0.6666666666666666; // 2/3
+    public static final float ROTATE3 = -0.16666666666666666f; // (1/sqrt(3) - 1) / 3
+    public static final float FALLBACK3 = 0.6666666666666666f; // 2/3
 
     // 3D: raio do kernel ao quadrado
-    public static final double KERNEL3 = 0.6;
+    public static final float KERNEL3 = 0.6f;
 
     // normalização 3D
-    public static final double NORM3   = 32.0;
+    public static final float NORM3   = 32.0f;
 
     // ===TABELAS DE GRADIENTE ===
     // 2D: 24 gradientes unitarios distribuidos simetricamente no circulo
@@ -112,18 +112,18 @@ public final class OpenSimplex2 {
      * retorna valor aproximadamente em [-1, 1]
      * kernel estendido(raio 3/4) para suavidade maior
      */
-    public static double ruido2(long semente, double x, double y) {
+    public static float ruido2(long semente, float x, float y) {
         // transforma para coordenadas do lattice simplex
-        double skew = (x + y) * SKEW2;
+        float skew = (x + y) * SKEW2;
         int i = floorRapido(x + skew);
         int j = floorRapido(y + skew);
 
         // coordenadas dentro da célula skewed
-        double unskew = (i + j) * UNSKEW2;
-        double x0 = x - (i - unskew);
-        double y0 = y - (j - unskew);
+        float unskew = (i + j) * UNSKEW2;
+        float x0 = x - (i - unskew);
+        float y0 = y - (j - unskew);
 
-        double valor = 0.0;
+        float valor = 0.0f;
 
         // contribuição dos vertices do triangulo
         // OpenSimplex2S avalia 4 vertices(não 3 como Simplex) pro kernel maior
@@ -131,25 +131,25 @@ public final class OpenSimplex2 {
         valor += contribuicao2(semente, i, j, x0, y0);
 
         // vertice 1 e 2: os dois da segunda metade do losango
-        double x1 = x0 - 1.0 + UNSKEW2;
-        double y1 = y0 + UNSKEW2;
-        double x2 = x0 + UNSKEW2;
-        double y2 = y0 - 1.0 + UNSKEW2;
+        float x1 = x0 - 1.0f + UNSKEW2;
+        float y1 = y0 + UNSKEW2;
+        float x2 = x0 + UNSKEW2;
+        float y2 = y0 - 1.0f + UNSKEW2;
 
         valor += contribuicao2(semente, i + 1, j,     x1, y1);
         valor += contribuicao2(semente, i,     j + 1, x2, y2);
 
         // vertice 3: canto oposto do losango
-        double x3 = x0 - 1.0 + 2.0 * UNSKEW2;
-        double y3 = y0 - 1.0 + 2.0 * UNSKEW2;
+        float x3 = x0 - 1.0f + 2.0f * UNSKEW2;
+        float y3 = y0 - 1.0f + 2.0f * UNSKEW2;
         valor += contribuicao2(semente, i + 1, j + 1, x3, y3);
 
         return valor * NORM2;
     }
 
-    public static double contribuicao2(long semente, int xi, int yi, double dx, double dy) {
-        double t = KERNEL2 - dx * dx - dy * dy;
-        if(t <= 0.0) return 0.0;
+    public static float contribuicao2(long semente, int xi, int yi, float dx, float dy) {
+        float t = KERNEL2 - dx * dx - dy * dy;
+        if(t <= 0.0) return 0.0f;
         t *= t;
         // indice no array de gradientes: hash -> [0, 24) -> *2 para pegar par(gx, gy)
         int gi = (hash(semente, xi, yi) & 23) * 2;
@@ -161,15 +161,15 @@ public final class OpenSimplex2 {
      * cortes horizontais(Y fixo) tem maxima isotropia
      * uso típico: ruido3XZ(semente, x * escala, y * escalaY, z * escala)
      */
-    public static double ruido3XZ(long semente, double x, double y, double z) {
+    public static float ruido3XZ(long semente, float x, float y, float z) {
         // rotação que alinha o eixo Y com a diagonal principal do BCC
         // e mantém XZ no plano de melhor isotropia
-        double xz = x + z;
-        double s2 = xz * ROTATE3;
-        double yr = y * FALLBACK3;
-        double xr = x + s2 + yr;
-        double zr = z + s2 + yr;
-        double yr2 = xz * (-FALLBACK3) + yr;
+        float xz = x + z;
+        float s2 = xz * ROTATE3;
+        float yr = y * FALLBACK3;
+        float xr = x + s2 + yr;
+        float zr = z + s2 + yr;
+        float yr2 = xz * (-FALLBACK3) + yr;
 
         return ruido3Base(semente, xr, yr2, zr);
     }
@@ -177,13 +177,13 @@ public final class OpenSimplex2 {
      * ruido 3D com planos XY primarios
      * util pra animações onde Z é o tempo, ou cavernas onde Y/Z são primarios
      */
-    public static double ruido3XY(long semente, double x, double y, double z) {
-        double xy = x + y;
-        double s2 = xy * ROTATE3;
-        double zr = z * FALLBACK3;
-        double xr = x + s2 + zr;
-        double yr = y + s2 + zr;
-        double zr2 = xy * (-FALLBACK3) + zr;
+    public static float ruido3XY(long semente, float x, float y, float z) {
+        float xy = x + y;
+        float s2 = xy * ROTATE3;
+        float zr = z * FALLBACK3;
+        float xr = x + s2 + zr;
+        float yr = y + s2 + zr;
+        float zr2 = xy * (-FALLBACK3) + zr;
 
         return ruido3Base(semente, xr, yr, zr2);
     }
@@ -195,24 +195,24 @@ public final class OpenSimplex2 {
      * total: 8 vertices contribuindo mais do que Simplex 3D(4 vertices)
      * resultado: muito mais isotropico, menos artefatos direcionais
      */
-    public static double ruido3Base(long semente, double xr, double yr, double zr) {
+    public static float ruido3Base(long semente, float xr, float yr, float zr) {
         // celula base no lattice
         int xrb = floorRapido(xr);
         int yrb = floorRapido(yr);
         int zrb = floorRapido(zr);
 
         // fração dentro da celula
-        double xri = xr - xrb;
-        double yri = yr - yrb;
-        double zri = zr - zrb;
+        float xri = xr - xrb;
+        float yri = yr - yrb;
+        float zri = zr - zrb;
 
         // determinamos quais dos 8 vertices do cubo contribuem
         // baseado na proximidade ao longo de cada eixo
-        double xNeg = xri - 0.5;
-        double yNeg = yri - 0.5;
-        double zNeg = zri - 0.5;
+        float xNeg = xri - 0.5f;
+        float yNeg = yri - 0.5f;
+        float zNeg = zri - 0.5f;
 
-        double valor = 0.0;
+        float valor = 0.0f;
 
         // primeiro half-lattice: 4 vertices
         valor += contribuicao3(semente, xrb, yrb, zrb, xri,  yri,  zri );
@@ -231,9 +231,9 @@ public final class OpenSimplex2 {
         return valor * NORM3;
     }
 
-    public static double contribuicao3(long semente, int xi, int yi, int zi, double dx, double dy, double dz) {
-        double t = KERNEL3 - dx * dx - dy * dy - dz * dz;
-        if(t <= 0.0) return 0.0;
+    public static float contribuicao3(long semente, int xi, int yi, int zi, float dx, float dy, float dz) {
+        float t = KERNEL3 - dx * dx - dy * dy - dz * dz;
+        if(t <= 0.0f) return 0.0f;
         t *= t;
         int gi = (hash(semente, xi, yi, zi) & 23) * 3;
         return t * t * (GRAD3[gi] * dx + GRAD3[gi + 1] * dy + GRAD3[gi + 2] * dz);
@@ -244,11 +244,11 @@ public final class OpenSimplex2 {
      * persistencia: quanto cada oitava contribui(tipicamente 0.5)
      * lacunaridade: quanto a frequencia aumenta por oitava(tipicamente 2.0)
      */
-    public static double ruido2Fractal(long semente, double x, double y, int oitavas, double persistencia, double lacunaridade) {
-        double total = 0.0;
-        double amplitude = 1.0;
-        double frequencia = 1.0;
-        double maximo = 0.0;
+    public static float ruido2Fractal(long semente, float x, float y, int oitavas, float persistencia, float lacunaridade) {
+        float total = 0.0f;
+        float amplitude = 1.0f;
+        float frequencia = 1.0f;
+        float maximo = 0.0f;
 
         for(int i = 0; i < oitavas; i++) {
             total += ruido2(semente, x * frequencia, y * frequencia) * amplitude;
@@ -262,11 +262,11 @@ public final class OpenSimplex2 {
     }
 	
     // fBm 3D XZ para terreno
-    public static double ruido3XZFractal(long semente, double x, double y, double z, int oitavas, double persistencia, double lacunaridade) {
-        double total = 0.0;
-        double amplitude = 1.0;
-        double frequencia = 1.0;
-        double maximo = 0.0;
+    public static float ruido3XZFractal(long semente, float x, float y, float z, int oitavas, float persistencia, float lacunaridade) {
+        float total = 0.0f;
+        float amplitude = 1.0f;
+        float frequencia = 1.0f;
+        float maximo = 0.0f;
 
         for(int i = 0; i < oitavas; i++) {
             total += ruido3XZ(semente, x * frequencia, y * frequencia, z * frequencia) * amplitude;
@@ -279,11 +279,11 @@ public final class OpenSimplex2 {
     }
 	
     // fBm 3D XY para cavernas e volumes
-    public static double ruido3XYFractal(long semente, double x, double y, double z, int oitavas, double persistencia, double lacunaridade) {
-        double total = 0.0;
-        double amplitude = 1.0;
-        double frequencia = 1.0;
-        double maximo = 0.0;
+    public static float ruido3XYFractal(long semente, float x, float y, float z, int oitavas, float persistencia, float lacunaridade) {
+        float total = 0.0f;
+        float amplitude = 1.0f;
+        float frequencia = 1.0f;
+        float maximo = 0.0f;
 
         for(int i = 0; i < oitavas; i++) {
             total += ruido3XY(semente, x * frequencia, y * frequencia, z * frequencia) * amplitude;
@@ -296,7 +296,7 @@ public final class OpenSimplex2 {
     }
 
 	// util:
-    public static int floorRapido(double x) {
+    public static int floorRapido(float x) {
         int xi = (int)x;
         return x < xi ? xi - 1 : xi;
     }
